Version2

+++ src/main/java/net/ktccenter/campusApi/dto/lite/administration/LiteJourOuvrableDTO.java
@@ -14,7 +14,7 @@ import net.ktccenter.campusApi.enums.Jour;
 public class LiteJourOuvrableDTO {
     private Long id;
     private Jour jour;
-    private Float[] intervalle;
+    private double[] intervalle;
 
     public LiteJourOuvrableDTO(JourOuvrable jourOuvrable) {
         this.id = jourOuvrable.getId();
